feat: Add custom headers support and content type updates for lamina
Custom header support has been added to the "lamina" decorator. Now, on returning a dictionary consisting of headers in the tuple, it gets merged with the existing headers. This will provide users with the ability to add custom values to the response headers.

In addition, updates were made to the content type handling. Content type can now be optionally defined for the decorated functions. Changes were made to the return statement of the response.

New tests were added to validate these changes and updates to the README were made to document the new feature and changes.